Restructure changelog module into submodules for greater readability
